[fix] 일지 작성 화면의 너비 별 UI 표시부분을 별도로 분리하여 가시성 개선
