+++ presentation/src/main/java/com/strayalphaca/presentation/screens/diary/write/component/block/PolaroidHorizontalPager.kt
@@ -1,4 +1,4 @@
-package com.strayalphaca.presentation.screens.diary.write.component
+package com.strayalphaca.presentation.screens.diary.write.component.block
 
 import android.net.Uri
 import androidx.compose.foundation.ExperimentalFoundationApi
